+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## 1.9.17 - 2026-02-27
+- Fixed **Camera Stand** not being affected by gravity.
 - [Forge] Removed dependency on the specific **Create** version.
 
 ## 1.9.16 - 2026-02-23